[mgsim-refactor] Re-enable compiler warnings and squash a few potential bugs found with them.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4461 e97e5017-a994-416e-809d-76780e9f78db

+++ arch/FPU.cpp
@@ -270,7 +270,7 @@ FPU::FPU(const std::string& name, Object& parent, Clock& clock, Config& config, 
         {
             throw InvalidArgumentException(*this, "NumUnits not set or zero");
         }
-        for (int i = 0; i < nUnits; ++i)
+        for (size_t i = 0; i < nUnits; ++i)
         {
             stringstream ssname;
             ssname << "Unit" << i;
@@ -292,7 +292,7 @@ FPU::FPU(const std::string& name, Object& parent, Clock& clock, Config& config, 
             }
             if (ops.empty())
             {
-                throw exceptf<InvalidArgumentException>(*this, "No operation specified for unit %d", i);
+                throw exceptf<InvalidArgumentException>(*this, "No operation specified for unit %zu", i);
             }
  
             // Add this unit into the mapping table for the ops it implements
